+++ lib/src/webdav/file.dart
@@ -57,14 +57,12 @@ List<WebDavFile> treeFromWebDavXml(String xmlStr) {
   for (final response in xmlDocument.findAllElements('d:response')) {
     final davItemName = response.findAllElements('d:href').single.text;
     final contentTypeElements = response.findAllElements('d:getcontenttype');
-    final contentType =
-        contentTypeElements.isNotEmpty && contentTypeElements.single.text != ''
-            ? contentTypeElements.single.text
-            : null;
+    final contentType = contentTypeElements.single.text != ''
+        ? contentTypeElements.single.text
+        : null;
     final contentLengthElements =
         response.findAllElements('d:getcontentlength');
-    final contentLength = contentLengthElements.isNotEmpty &&
-            contentLengthElements.single.text != ''
+    final contentLength = contentLengthElements.single.text != ''
         ? int.parse(contentLengthElements.single.text)
         : 0;